requirements.txt
# Bird/Plane/Superman Image Classifier Dependencies

# Core deep learning
torch>=2.0.0
torchvision>=0.15.0

# Image processing
Pillow>=9.0.0
imagehash>=4.3.0

# Data science
numpy>=1.21.0
scikit-learn>=1.0.0

# Progress bars and utilities
tqdm>=4.65.0

# Image downloading (for dataset creation)
duckduckgo-search>=3.8.0
requests>=2.28.0

# Optional: for visualization
matplotlib>=3.5.0
